remove unneeded pages and migrate to expo navigation
