+++ frontend/app/Activity.jsx
