+++ frontend/app/index.jsx
@@ -8,6 +8,7 @@ import {
 } from "react-native";
 import { BleManager } from "react-native-ble-plx";
 import LocalHost from "../components/data/LocalHost";
+import { router } from 'expo-router';
 
 const Connect = ({ exitConnect }) => {
   const ipAddress = LocalHost.ipAddress;
@@ -95,6 +96,7 @@ const Connect = ({ exitConnect }) => {
       }
     }
   };
+
   // allows developers to exit connect screen without connecting to a device
   const handleDevExit = async () => {
     exitConnect(deviceIndex);
@@ -115,7 +117,7 @@ const Connect = ({ exitConnect }) => {
             borderRadius: 5,
           },
         ]}
-        onPress={handleDevExit}
+        onPress={() => router.push("/Monitor")}
       >
         <Text style={[styles.text, { color: "white" }]}>Dev Mode</Text>
       </TouchableOpacity>
