+++ frontend/components/ActivityCalculations.js
