+++ frontend/components/MonitorCalculations.js
