+++ frontend/components/SlidingButton.jsx
