+++ frontend/package.json
@@ -19,7 +19,7 @@
     "axios": "^1.6.7",
     "babel-plugin-macros": "^3.1.0",
     "expo": "^50.0.7",
-    "expo-status-bar": "^1.11.1",
+    "expo-status-bar": "~1.11.1",
     "expo-system-ui": "^2.9.3",
     "metro-runtime": "^0.80.6",
     "react": "18.2.0",
@@ -31,7 +31,12 @@
     "react-native-public-ip": "^1.0.2",
     "react-native-svg": "^15.1.0",
     "react-native-swiper": "^1.6.0",
-    "styled-components": "^6.1.8"
+    "styled-components": "^6.1.8",
+    "expo-router": "~3.4.8",
+    "react-native-safe-area-context": "4.8.2",
+    "react-native-screens": "~3.29.0",
+    "expo-linking": "~6.2.2",
+    "expo-constants": "~15.4.5"
   },
   "devDependencies": {
     "@babel/core": "^7.24.0",
